rename the command to "psql"
